🐛fix: 필터링된 이슈들의 개수만 세도록 수정

+++ backend/src/main/java/codesquard/app/issue/mapper/response/IssueFilterResponse.java
@@ -18,10 +18,10 @@ public class IssueFilterResponse {
 	private String input;
 
 	@JsonProperty("openedIssueCount")
-	private Long openedIssueCount;
+	private int openedIssueCount;
 
 	@JsonProperty("closedIssueCount")
-	private Long closedIssueCount;
+	private int closedIssueCount;
 
 	@JsonProperty("labelCount")
 	private Long labelCount;
